feat: Rename `getFixturesEngine` to `setupFixturesEngine`
BREAKING CHANGE:
`get-fixture-engine.js` util was renamed to `setup-fixture-engine.js`
